@@ -16,77 +16,82 @@
         "ios": "ionic cordova build ios"
     },
     "dependencies": {
-        "@angular/common": "4.3.3",
-        "@angular/compiler": "4.3.3",
-        "@angular/compiler-cli": "4.3.3",
-        "@angular/core": "4.3.3",
-        "@angular/forms": "4.3.3",
-        "@angular/http": "4.3.3",
-        "@angular/platform-browser": "4.3.3",
-        "@angular/platform-browser-dynamic": "4.3.3",
-        "@angular/platform-server": "4.3.3",
-        "@angular/tsc-wrapped": "4.3.3",
-        "@ionic-native/android-permissions": "4.2.0",
-        "@ionic-native/barcode-scanner": "4.1.0",
-        "@ionic-native/camera": "4.1.0",
-        "@ionic-native/core": "4.1.0",
-        "@ionic-native/device-feedback": "4.1.0",
-        "@ionic-native/facebook": "4.1.0",
-        "@ionic-native/file": "4.1.0",
-        "@ionic-native/file-chooser": "4.1.0",
-        "@ionic-native/file-path": "4.1.0",
-        "@ionic-native/native-storage": "4.1.0",
-        "@ionic-native/network": "4.2.0",
-        "@ionic-native/splash-screen": "4.1.0",
-        "@ionic-native/status-bar": "4.1.0",
+        "@angular/common": "4.4.0-RC.0",
+        "@angular/compiler": "4.4.0-RC.0",
+        "@angular/compiler-cli": "4.4.0-RC.0",
+        "@angular/core": "4.4.0-RC.0",
+        "@angular/forms": "4.4.0-RC.0",
+        "@angular/http": "4.4.0-RC.0",
+        "@angular/platform-browser": "4.4.0-RC.0",
+        "@angular/platform-browser-dynamic": "4.4.0-RC.0",
+        "@angular/platform-server": "4.4.0-RC.0",
+        "@angular/tsc-wrapped": "4.4.0-RC.0",
+        "@ionic-native/android-permissions": "4.2.1",
+        "@ionic-native/barcode-scanner": "4.2.1",
+        "@ionic-native/camera": "4.2.1",
+        "@ionic-native/core": "4.2.1",
+        "@ionic-native/device-feedback": "4.2.1",
+        "@ionic-native/facebook": "4.2.1",
+        "@ionic-native/file": "4.2.1",
+        "@ionic-native/file-chooser": "4.2.1",
+        "@ionic-native/file-path": "4.2.1",
+        "@ionic-native/native-storage": "4.2.1",
+        "@ionic-native/network": "4.2.1",
+        "@ionic-native/splash-screen": "4.2.1",
+        "@ionic-native/status-bar": "4.2.1",
         "@ionic-native/transfer": "3.14.0",
         "@ionic/storage": "2.0.1",
-        "@ngx-translate/core": "7.1.0",
-        "@ngx-translate/http-loader": "1.0.1",
+        "@ngx-translate/core": "8.0.0",
+        "@ngx-translate/http-loader": "1.0.2",
         "angular2-uuid": "1.1.1",
-        "angularfire2": "4.0.0-rc.1",
-        "chart.js": "2.6.0",
-        "com.megster.cordova.FileChooser": "http://github.com/don/cordova-filechooser.git",
-        "cordova-android": "~6.2.3",
+        "angularfire2": "4.0.0-rc.2",
+        "chart.js": "2.7.0",
+        "com.megster.cordova.FileChooser": "https://github.com/loicsalou/cordova-filechooser.git",
+        "cordova-android": "6.2.3",
         "cordova-plugin-android-permissions": "1.0.0",
-        "cordova-plugin-console": "~1.0.7",
-        "cordova-plugin-device": "~1.1.6",
-        "cordova-plugin-file": "~4.3.3",
-        "cordova-plugin-filepath": "~1.0.2",
+        "cordova-plugin-camera": "2.4.1",
+        "cordova-plugin-compat": "1.1.0",
+        "cordova-plugin-console": "1.0.7",
+        "cordova-plugin-device": "1.1.6",
+        "cordova-plugin-facebook4.4": "1.0.1",
+        "cordova-plugin-file": "4.3.3",
+        "cordova-plugin-file-chooser": "git+https://github.com/loicsalou/cordova-filechooser.git",
+        "cordova-plugin-filechooser": "1.0.1",
+        "cordova-plugin-filepath": "1.0.2",
+        "cordova-plugin-image-picker": "1.1.3",
         "cordova-plugin-nativestorage": "2.2.2",
         "cordova-plugin-network-information": "1.3.3",
-        "cordova-plugin-splashscreen": "~4.0.3",
-        "cordova-plugin-statusbar": "~2.2.3",
+        "cordova-plugin-splashscreen": "4.0.3",
+        "cordova-plugin-statusbar": "2.2.3",
         "cordova-plugin-velda-devicefeedback": "0.0.2",
-        "cordova-plugin-whitelist": "~1.3.2",
-        "firebase": "4.2.0",
-        "firebase-rxjs": "0.1.1",
-        "firebase-rxjs-angular": "0.1.1",
-        "ionic-angular": "3.6.0",
-        "ionic-plugin-keyboard": "~2.2.1",
+        "cordova-plugin-whitelist": "1.3.2",
+        "firebase": "4.3.1",
+        "firebase-rxjs": "0.1.2",
+        "firebase-rxjs-angular": "0.1.2",
+        "ionic-angular": "3.6.1",
+        "ionic-plugin-keyboard": "2.2.1",
         "ionicons": "3.0.0",
         "lodash": "4.17.4",
         "ng2-charts": "1.6.0",
         "phonegap-plugin-barcodescanner": "6.0.8",
-        "rxjs": "5.4.2",
+        "rxjs": "5.4.3",
         "sw-toolbox": "3.6.0",
         "text-encoding": "0.6.4",
-        "zone.js": "0.8.16"
+        "zone.js": "0.8.17"
     },
     "devDependencies": {
-        "@angular/cli": "1.2.7",
-        "@ionic/app-scripts": "2.1.3",
-        "@types/jasmine": "2.5.53",
+        "@angular/cli": "1.4.1",
+        "@ionic/app-scripts": "2.1.4",
+        "@types/jasmine": "2.5.54",
         "@types/karma": "0.13.36",
-        "@types/node": "8.0.19",
-        "ionic": "3.9.2",
-        "jasmine-core": "2.7.0",
-        "jasmine-spec-reporter": "4.1.1",
+        "@types/node": "8.0.28",
+        "ionic": "3.10.3",
+        "jasmine-core": "2.8.0",
+        "jasmine-spec-reporter": "4.2.1",
         "promise-polyfill": "6.0.2",
-        "typescript": "2.4.2"
+        "typescript": "2.5.2"
     },
     "cordovaPlugins": [
-        "com.megster.cordova.FileChooser",
         "cordova-plugin-camera",
         "cordova-plugin-compat",
         "cordova-plugin-console",
